Integrate Pendo tracking script
