+++ package.json
@@ -20,11 +20,11 @@
     "pre-push": "yarn build"
   },
   "dependencies": {
-    "@docsearch/react": "^3.3.0",
-    "@edgeandnode/components": "^24.0.7",
+    "@docsearch/react": "^3.3.2",
+    "@edgeandnode/components": "^24.5.0",
     "@emotion/react": "^11.10.5",
-    "@mdx-js/loader": "^2.1.5",
-    "@mdx-js/react": "^2.1.5",
+    "@mdx-js/loader": "^2.2.1",
+    "@mdx-js/react": "^2.2.1",
     "@next/mdx": "^12.3.4",
     "@radix-ui/react-collapsible": "1.0.1",
     "@radix-ui/react-popover": "^1.0.2",
@@ -32,7 +32,7 @@
     "@reach/auto-id": "^0.18.0",
     "lodash": "^4.17.21",
     "next": "^12.3.4",
-    "next-seo": "^5.14.1",
+    "next-seo": "^5.15.0",
     "prism-react-renderer": "^1.3.5",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -48,20 +48,20 @@
     "@svgr/webpack": "^6.5.1",
     "@types/color": "^3.0.3",
     "@types/lodash": "^4.14.191",
-    "@types/node": "^16.18.4",
+    "@types/node": "^16.18.11",
     "@types/react": "^17.0.52",
     "@types/react-dom": "^17.0.18",
     "acorn": "^8.8.1",
     "acorn-jsx": "^5.3.2",
-    "eslint": "^8.28.0",
-    "husky": "^8.0.2",
-    "lint-staged": "^13.0.4",
+    "eslint": "^8.31.0",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.1.0",
     "path": "^0.12.7",
-    "prettier": "^2.8.0",
+    "prettier": "^2.8.2",
     "remark-frontmatter": "^4.0.1",
     "remark-mdx-frontmatter": "^2.1.1",
     "serialize-as-code": "^2.0.2",
-    "typescript": "4.9.3",
+    "typescript": "4.9.4",
     "unist-util-visit": "^4.1.1",
     "url": "^0.11.0"
   },
